third_party/blink/web_tests/http/tests/mojo/versioned-optional-numerics.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharness-helpers.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type="module">
import {
  InterfaceV2Remote,
  InterfaceV2Receiver,
  RegularEnum,
} from "/gen/content/web_test/common/mojo_optional_numerics_unittest.mojom.m.js"

class InterfaceV2Impl {
  constructor() {
    this.receiver = new InterfaceV2Receiver(this);
  }

  async methodWithVersionedParams(
    boolValue,
    uint8Value,
    int8Value,
    uint16Value,
    int16Value,
    uint32Value,
    int32Value,
    uint64Value,
    int64Value,
    floatValue,
    doubleValue,
    enumValue,
  ) {
    return {
      boolValue: boolValue === null ? null : !boolValue,
      uint8Value: uint8Value === null ? null : uint8Value * 2,
      int8Value: int8Value === null ? null : int8Value * 2,
      uint16Value: uint16Value === null ? null : uint16Value * 2,
      int16Value: int16Value === null ? null : int16Value * 2,
      uint32Value: uint32Value === null ? null : uint32Value * 2,
      int32Value: int32Value === null ? null : int32Value * 2,
      uint64Value: uint64Value === null ? null : uint64Value * BigInt('2'),
      int64Value: int64Value === null ? null : int64Value * BigInt('2'),
      floatValue: floatValue === null ? null : floatValue * 3,
      doubleValue: doubleValue === null ? null : doubleValue * 3,
      enumValue: enumValue === null ? null : enumValue - 1,
    };
  }
}

const impl = new InterfaceV2Impl;
const remote = impl.receiver.$.bindNewPipeAndPassRemote();

promise_test(async () => {
  const nullArgs = {
    boolValue: null,
    uint8Value: null,
    int8Value: null,
    uint16Value: null,
    int16Value: null,
    uint32Value: null,
    int32Value: null,
    uint64Value: null,
    int64Value: null,
    floatValue: null,
    doubleValue: null,
    enumValue: null,
  };

  const expectedResponse = {
    boolValue: null,
    uint8Value: null,
    int8Value: null,
    uint16Value: null,
    int16Value: null,
    uint32Value: null,
    int32Value: null,
    uint64Value: null,
    int64Value: null,
    floatValue: null,
    doubleValue: null,
    enumValue: null,
  };

  const nullArgsResponse =
    await remote.methodWithVersionedParams(...Object.values(nullArgs));
  assert_weak_equals(nullArgsResponse, expectedResponse);

  const undefinedArgs = {
    boolValue: undefined,
    uint8Value: undefined,
    int8Value: undefined,
    uint16Value: undefined,
    int16Value: undefined,
    uint32Value: undefined,
    int32Value: undefined,
    uint64Value: undefined,
    int64Value: undefined,
    floatValue: undefined,
    doubleValue: undefined,
    enumValue: undefined,
  };

  const undefinedArgsResponse =
    await remote.methodWithVersionedParams(...Object.values(undefinedArgs));
  assert_weak_equals(undefinedArgsResponse, expectedResponse);
}, 'Basic test for sending/receiving null numerics in versioned ' +
   'interfaces.');

promise_test(async () => {
  const args = {
    boolValue: true,
    uint8Value: 8,
    int8Value: -8,
    uint16Value: 16,
    int16Value: -16,
    uint32Value: 32,
    int32Value: -32,
    uint64Value: BigInt('64'),
    int64Value: BigInt('-64'),
    floatValue: -0.5,
    doubleValue: 0.25,
    enumValue: RegularEnum.kBar,
  };
  const expectedResponse = {
    boolValue: !args.boolValue,
    uint8Value: 16,
    int8Value: -16,
    uint16Value: 32,
    int16Value: -32,
    uint32Value: 64,
    int32Value: -64,
    uint64Value: BigInt('128'),
    int64Value: BigInt('-128'),
    floatValue: -1.5,
    doubleValue: 0.75,
    enumValue: RegularEnum.kFoo,
  };

  const response =
    await remote.methodWithVersionedParams(...Object.values(args));
  assert_weak_equals(response, expectedResponse);
}, 'Basic test for sending/receiving optional numerics in versioned ' +
   'interfaces.');
</script>
